Menzilim/menzilim-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.19
httpx==0.28.1
redis==5.2.1
pillow==11.0.0
aiofiles==24.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
